# openapi.json
{
  "components": {
    "schemas": {
...
      "DigestOut": {
        "properties": {
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "DigestOut",
        "type": "object"
      },
...
      "FindItem": {
        "properties": {
          "camera": {
            "title": "Camera",
            "type": "string"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "similarity": {
            "title": "Similarity",
            "type": "number"
          },
          "time": {
            "format": "date-time",
            "title": "Time",
            "type": "string"
          }
        },
        "required": [
          "time",
          "camera",
          "similarity",
          "photo_url"
        ],
        "title": "FindItem",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HourBucket": {
        "properties": {
          "alerts": {
            "title": "Alerts",
            "type": "integer"
          },
          "events": {
            "title": "Events",
            "type": "integer"
          },
          "hour": {
            "title": "Hour",
            "type": "string"
          }
        },
        "required": [
          "hour",
          "events",
          "alerts"
        ],
        "title": "HourBucket",
        "type": "object"
      },
...
      "OkOut": {
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "OkOut",
        "type": "object"
      },
      "OverviewOut": {
        "properties": {
          "alerts_today": {
            "title": "Alerts Today",
            "type": "integer"
          },
          "cameras_online": {
            "title": "Cameras Online",
            "type": "integer"
          },
          "cameras_total": {
            "title": "Cameras Total",
            "type": "integer"
          },
          "false_positives_today": {
            "title": "False Positives Today",
            "type": "integer"
          },
          "hourly": {
            "items": {
              "$ref": "#/components/schemas/HourBucket"
            },
            "title": "Hourly",
            "type": "array"
          },
          "recent_alerts": {
            "items": {
              "$ref": "#/components/schemas/EventOut"
            },
            "title": "Recent Alerts",
            "type": "array"
          },
          "visitors_today": {
            "title": "Visitors Today",
            "type": "integer"
          }
        },
        "required": [
          "visitors_today",
          "alerts_today",
          "cameras_online",
          "cameras_total",
          "false_positives_today",
          "hourly",
          "recent_alerts"
        ],
        "title": "OverviewOut",
        "type": "object"
      },
      "PasswordChange": {
        "properties": {
          "new_password": {
            "title": "New Password",
            "type": "string"
          }
        },
        "required": [
          "new_password"
        ],
        "title": "PasswordChange",
        "type": "object"
      },
      "PersonAppearance": {
        "properties": {
          "camera": {
            "title": "Camera",
            "type": "string"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "time": {
            "format": "date-time",
            "title": "Time",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "time",
          "camera",
          "type",
          "photo_url"
        ],
        "title": "PersonAppearance",
        "type": "object"
      },
...
      "StatsOut": {
        "properties": {
          "alerts": {
            "title": "Alerts",
            "type": "integer"
          },
          "by_camera": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Camera",
            "type": "object"
          },
          "by_type": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Type",
            "type": "object"
          },
          "false_positives": {
            "title": "False Positives",
            "type": "integer"
          },
          "hours": {
            "title": "Hours",
            "type": "number"
          },
          "visitors_entered": {
            "title": "Visitors Entered",
            "type": "integer"
          }
        },
        "required": [
          "hours",
          "visitors_entered",
          "alerts",
          "false_positives",
          "by_type",
          "by_camera"
        ],
        "title": "StatsOut",
        "type": "object"
      },
...
      "SystemOut": {
        "properties": {
          "agent_enabled": {
            "title": "Agent Enabled",
            "type": "boolean"
          },
          "cameras": {
            "items": {
              "$ref": "#/components/schemas/SystemCamera"
            },
            "title": "Cameras",
            "type": "array"
          },
          "events_total": {
            "title": "Events Total",
            "type": "integer"
          },
          "face_recognition": {
            "title": "Face Recognition",
            "type": "boolean"
          },
          "media_dir": {
            "title": "Media Dir",
            "type": "string"
          },
          "offline_mode": {
            "title": "Offline Mode",
            "type": "boolean"
          },
          "site_name": {
            "title": "Site Name",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "vehicle_detection": {
            "title": "Vehicle Detection",
            "type": "boolean"
          },
          "vlm_enabled": {
            "title": "Vlm Enabled",
            "type": "boolean"
          },
          "weapon_detection": {
            "title": "Weapon Detection",
            "type": "boolean"
          },
          "yolo_model": {
            "title": "Yolo Model",
            "type": "string"
          }
        },
        "required": [
          "site_name",
          "timezone",
          "cameras",
          "yolo_model",
          "agent_enabled",
          "vlm_enabled",
          "face_recognition",
          "weapon_detection",
          "vehicle_detection",
          "offline_mode",
          "media_dir",
          "events_total"
        ],
        "title": "SystemOut",
        "type": "object"
      },
...
      "TotpCode": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "TotpCode",
        "type": "object"
      },
      "TotpSetupOut": {
        "properties": {
          "otpauth_uri": {
            "title": "Otpauth Uri",
            "type": "string"
          },
          "qr_svg": {
            "title": "Qr Svg",
            "type": "string"
          },
          "secret": {
            "title": "Secret",
            "type": "string"
          }
        },
        "required": [
          "secret",
          "otpauth_uri",
          "qr_svg"
        ],
        "title": "TotpSetupOut",
        "type": "object"
      },
      "TotpStatus": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "TotpStatus",
        "type": "object"
      },
      "UserCreate": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "password": {
            "title": "Password",
            "type": "string"
          },
          "role": {
            "default": "user",
            "title": "Role",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "UserCreate",
        "type": "object"
      },
      "UserOut": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "last_login_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "locked": {
            "title": "Locked",
            "type": "boolean"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "role",
          "is_active",
          "full_name",
          "email",
          "last_login_at",
          "locked"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "UserPatch": {
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          }
        },
        "title": "UserPatch",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI Security Copilot — dashboard backend",
    "title": "Pragmata AI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/digest": {
      "get": {
        "operationId": "digest_api_v1_digest_get",
        "parameters": [
          {
            "in": "query",
            "name": "hours",
            "required": false,
            "schema": {
              "default": 24,
              "exclusiveMinimum": 0,
              "maximum": 168,
              "title": "Hours",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "ru",
              "enum": [
                "ru",
                "uz",
                "en"
              ],
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DigestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Digest",
        "tags": [
          "insights"
        ]
      }
    },
...
    "/api/v1/find": {
      "get": {
        "operationId": "find_api_v1_find_get",
        "parameters": [
          {
            "in": "query",
            "name": "description",
            "required": true,
            "schema": {
              "minLength": 3,
              "title": "Description",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hours",
            "required": false,
            "schema": {
              "default": 48,
              "exclusiveMinimum": 0,
              "maximum": 336,
              "title": "Hours",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FindItem"
                  },
                  "title": "Response Find Api V1 Find Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Find",
        "tags": [
          "insights"
        ]
      }
    },
...
    "/api/v1/me/2fa": {
      "get": {
        "operationId": "totp_status_api_v1_me_2fa_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TotpStatus"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Totp Status",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/me/2fa/disable": {
      "post": {
        "operationId": "totp_disable_api_v1_me_2fa_disable_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpCode"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Totp Disable",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/me/2fa/enable": {
      "post": {
        "operationId": "totp_enable_api_v1_me_2fa_enable_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpCode"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Totp Enable",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/me/2fa/setup": {
      "post": {
        "operationId": "totp_setup_api_v1_me_2fa_setup_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TotpSetupOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Totp Setup",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/me/password": {
      "post": {
        "operationId": "change_own_password_api_v1_me_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change Own Password",
        "tags": [
          "users"
        ]
      }
    },
...
    "/api/v1/overview": {
      "get": {
        "operationId": "overview_api_v1_overview_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Overview",
        "tags": [
          "insights"
        ]
      }
    },
...
    "/api/v1/stats": {
      "get": {
        "operationId": "stats_api_v1_stats_get",
        "parameters": [
          {
            "in": "query",
            "name": "hours",
            "required": false,
            "schema": {
              "default": 24,
              "exclusiveMinimum": 0,
              "maximum": 720,
              "title": "Hours",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Stats",
        "tags": [
          "insights"
        ]
      }
    },
    "/api/v1/system": {
      "get": {
        "operationId": "system_api_v1_system_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "System",
        "tags": [
          "insights"
        ]
      }
    },
    "/api/v1/tracks/{track_id}/photo": {
      "get": {
        "operationId": "track_photo_api_v1_tracks__track_id__photo_get",
        "parameters": [
          {
            "in": "path",
            "name": "track_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Track Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Track Photo",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/tracks/{track_id}/timeline": {
      "get": {
        "operationId": "timeline_api_v1_tracks__track_id__timeline_get",
        "parameters": [
          {
            "in": "path",
            "name": "track_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Track Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hours",
            "required": false,
            "schema": {
              "default": 24,
              "exclusiveMinimum": 0,
              "maximum": 168,
              "title": "Hours",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PersonAppearance"
                  },
                  "title": "Response Timeline Api V1 Tracks  Track Id  Timeline Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Timeline",
        "tags": [
          "insights"
        ]
      }
    },
    "/api/v1/users": {
      "get": {
        "operationId": "list_users_api_v1_users_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserOut"
                  },
                  "title": "Response List Users Api V1 Users Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Users",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "create_user_api_v1_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create User Api V1 Users Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create User",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{user_id}": {
      "patch": {
        "operationId": "patch_user_api_v1_users__user_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch User",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{user_id}/password": {
      "post": {
        "operationId": "reset_password_api_v1_users__user_id__password_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reset Password",
        "tags": [
          "users"
        ]
      }
    },
...
  }
}